# Conflicts:
#	logs/error.log
#	logs/log.2019-05-08.log
#	logs/log.log
#	src/main/resources/mapper/MeterMapper.xml
#	src/main/resources/mapper/UserMapper.xml

+++ src/main/java/com/iot/smart/water/meter/controller/MeterController.java
@@ -40,7 +40,7 @@ public class MeterController {
 	@PostMapping("/setMemberVolume")
 	@CrossOrigin(origins="*")
     public Response setMemberVolume(@RequestParam("memberName") String memberName,
-                                    @RequestParam("volume") float volume) {
+                                    @RequestParam("volume") long volume) {
         Response response = new Response();
         Meter meter = mService.getMeterByName(memberName);
         if (meter == null || volume <= 0) {
