+++ src/main/java/com/iot/smart/water/meter/service/MeterService.java
@@ -8,7 +8,7 @@ public interface MeterService {
 
     List<WaterBill> getWaterBill();
 
-    boolean setMemberVolume(Meter meter, float volume);
+    boolean setMemberVolume(Meter meter, long volume);
 
     Meter getMeterByName(String memberName);
